[GREEN] Simplify map layer management by introducing MapChartManager for better maintainability and modularity; Fixed map background not changing; Fixed non-Indonesian region being colored
