+++ app/components/floating_buttons/HumidityButton.tsx
@@ -40,7 +40,7 @@ export default function HumidityButton({
     if (mapService) {
       /* istanbul ignore next */
       if (isActive) mapService.showHumidityLayer()
-      else mapService.hideHumidityLayer()
+      // else mapService.hideHumidityLayer()
     }
   }, [isActive, mapService])
