+++ app/components/floating_buttons/SeverityButton.tsx
@@ -41,10 +41,11 @@ export default function GearToggleButton({
 
   // Add useEffect to respond to activeButton changes
   useEffect(() => {
-    if (isActive) {
-      mapService?.showSeverityLayer();
-    } else {
-      mapService?.hideSeverityLayer();
+    /* istanbul ignore next */
+    if (mapService) {
+      /* istanbul ignore next */
+      if (isActive) mapService.showSeverityLayer()
+      // else mapService.hideSeverityLayer()
     }
   }, [isActive, mapService])
