+++ app/components/floating_buttons/TemperatureButton.tsx
@@ -41,7 +41,7 @@ export default function TemperatureButton({
     if (mapService) {
       /* istanbul ignore next */
       if (isActive) mapService.showTemperatureLayer() 
-      else mapService.hideTemperatureLayer()
+      // else mapService.hideTemperatureLayer()
     }
   }, [isActive, mapService])
